[주영은] :sparkles: :bug: [feat & fix] 댓글 기능 및 부스 코드 수정

+++ services/boothService.js
@@ -1,4 +1,5 @@
 import boothRepository from "../repositorys/boothRepository.js";
+import participationRepository from "../repositorys/participationRepository.js";
 
 const createBooth = async (userId, festivalId, data) => {
   const booth = await boothRepository.createBooth(userId, festivalId, data);
@@ -6,13 +7,26 @@ const createBooth = async (userId, festivalId, data) => {
 };
 
 const getBoothAdmin = async (
+  adminId,
   festivalId,
   page,
   pageSize,
   orderBy,
   keyword,
-  type
+  type,
+  userRole
 ) => {
+  if (userRole !== "ADMIN") {
+    throw new Error("ADMIN 권한만 조회할 수 있습니다.");
+  }
+
+  const isParticipated = await participationRepository.participationCheck(
+    adminId,
+    festivalId
+  );
+  if (!isParticipated) {
+    throw new Error("참여중인 축제가 아닙니다.");
+  }
   return await boothRepository.getBoothAdmin(
     festivalId,
     page,
@@ -24,6 +38,7 @@ const getBoothAdmin = async (
 };
 
 const getBooths = async (
+  userId,
   festivalId,
   page,
   pageSize,
@@ -31,6 +46,14 @@ const getBooths = async (
   keyword,
   type
 ) => {
+  const isParticipated = await participationRepository.participationCheck(
+    userId,
+    festivalId
+  );
+  if (!isParticipated) {
+    throw new Error("참여중인 축제가 아닙니다.");
+  }
+
   return await boothRepository.getBooths(
     festivalId,
     page,
@@ -41,24 +64,64 @@ const getBooths = async (
   );
 };
 
-const getBooth = async (boothId) => {
+const getBooth = async (boothId, userId, festivalId) => {
+  const isParticipated = await participationRepository.participationCheck(
+    userId,
+    festivalId
+  );
+  if (!isParticipated) {
+    throw new Error("참여중인 축제가 아닙니다.");
+  }
   return await boothRepository.getBooth(boothId);
 };
 
-const updateBooth = async (boothId, userId, data) => {
-  const boothUser = await boothRepository.getMyBooths(userId);
-
-  if (boothUser[0].userId !== userId) {
-    throw new Error("해당 부스 소유자만 수정할 수 있습니다.");
+const updateBooth = async (festivalId, boothId, userId, userRole, data) => {
+  if (userRole !== "ADMIN") {
+    const boothUser = await boothRepository.getBooth(boothId);
+    if (boothUser.length === 0) {
+      throw new Error("수정할 부스가 없습니다.");
+    }
+    if (boothUser.userId !== userId) {
+      throw new Error("본인의 부스만 수정할 수 있습니다.");
+    }
   }
+
+  const isParticipated = await participationRepository.participationCheck(
+    userId,
+    festivalId
+  );
+  if (!isParticipated) {
+    throw new Error("참여중인 축제가 아닙니다.");
+  }
+
   return await boothRepository.updateBooth(boothId, data);
 };
 
 const getMyBooths = async (userId, festivalId) => {
+  const isParticipated = await participationRepository.participationCheck(
+    userId,
+    festivalId
+  );
+  if (!isParticipated) {
+    throw new Error("참여중인 축제가 아닙니다.");
+  }
   return await boothRepository.getMyBooths(userId, festivalId);
 };
 
-const deleteBooth = async (boothId) => {
+const deleteBooth = async (boothId, userId, festivalId, userRole) => {
+  const isParticipated = await participationRepository.participationCheck(
+    userId,
+    festivalId
+  );
+
+  if (!isParticipated) {
+    throw new Error("참여중인 축제가 아닙니다.");
+  }
+
+  const booth = await boothRepository.getBooth(parseInt(boothId));
+  if (booth.userId !== userId && userRole !== "ADMIN") {
+    throw new Error("삭제 권한이 없습니다.");
+  }
   return await boothRepository.deleteBooth(boothId);
 };
